refactor(shell): split output buffer helpers
